refactor(backup/s3): move utils to dedicated package
